[Move `@kbn/config-schema` to server] `maps_ems`
