+++ src/plugins/maps_ems/server/index.ts
@@ -7,7 +7,7 @@
  */
 
 import { PluginInitializerContext, PluginConfigDescriptor } from '@kbn/core/server';
-import { MapConfig, mapConfigSchema } from '../config';
+import { MapConfig, mapConfigSchema } from './config';
 export type { EMSSettings } from '../common';
 export type { MapsEmsPluginServerSetup } from './plugin';
